+++ src/domain/ingredients/ingredients.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
-import { Model } from 'mongoose';
+import { Model, Types } from 'mongoose';
 
 import { Ingredient, IngredientDocument } from './ingredients.schema';
 
@@ -9,8 +9,9 @@ import { GetIngredientByIdDto } from './dto/get-ingredient-by-id.dto';
 import { GetIngredientsDto } from './dto/getAlll.dto';
 import { UpdateIngredientDto } from './dto/update-ingredient.dto';
 import { CocktailsService } from '../cocktails/cocktails.service';
-import { Types } from 'mongoose';
-
+//
+import errorGenerator from '../../helpers/errorGenerator';
+//
 @Injectable()
 export class IngredientsService {
   constructor(
@@ -42,19 +43,22 @@ export class IngredientsService {
       '-owner',
     );
 
-    return newIngredient;
+    const defaultIngredients = await this.getDefault();
+    return newIngredient.concat(defaultIngredients);
   }
 
   async getIngredientById({ id }: GetIngredientByIdDto) {
     const ingredient = await this.ingredientModel.findById(id);
-
+    if (!ingredient) {
+      errorGenerator('Wrong ID , ingredient not found', 'NOT_FOUND');
+    }
     ingredient.cocktails = await this.cocktailsService.findByIngredient({ id });
 
     return ingredient.populate('cocktails.ingredients.data', ['id', 'title']);
   }
 
-  async deleteIngredient({ id }): Promise<void> {
-    await this.ingredientModel.findOneAndDelete({ id });
+  async deleteIngredient({ id, owner }): Promise<void> {
+    await this.ingredientModel.findOneAndDelete({ _id: id, owner });
   }
 
   async updateIngredient(
